ai_processing/resume_openai/requirements.txt
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
pytest>=7.0.0 